pyproject.toml
[build-system]
requires = [
    "setuptools >=65.0",
    "wheel >=0.36.2",
]
build-backend = "setuptools.build_meta"

[project]
name = "app"
authors = [{name = "Anthony Corletti", email = "anthcor@gmail.com"}]
dynamic = ["version", "readme"]
description = "app"
requires-python = ">=3.10"
dependencies = [
    "modal-client >=0.1.0",
    "fastapi >=0.70.0",
    "python-dotenv >=0.21.0",
    "structlog >= 21.2.0",
]

[project.optional-dependencies]
test = [
    "pytest >=6.2.5",
    "coverage >=6.1.1",
    "pytest-cov >=3.0.0",
]
dev = [
    "uvicorn >=0.15.0",
    "mypy >=0.910",
    "black >=21.10b0",
    "isort >=5.9.3",
    "ruff >=0.0.98",
    "pre-commit >=2.17.0",
]

[tool.setuptools]
license-files = ["LICENSE"]

[tool.setuptools.dynamic]
version = {attr = "app.__version__"}
readme = {file = "README.md", encoding = "utf-8"}

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 88
ignore = [
    "D10",
    "D415",
]
target-version = "py310"
extend-select = ["D"]
extend-ignore = [
    "D203",
    "D204",
    "D213",
    "D215",
    "D400",
    "D404",
    "D406",
    "D407",
    "D408",
    "D409",
    "D413",
]
